Treat 'self' and 'module()object' differently
Previously any call that began with 'self' or 'module()object' was handled by searching for type '$elf' from bottom of the scope.
Realized that 'module()object' actually refers to the enclosing module, not just the closest scope.
Added new 'find' methods to scope that allow for searching from the top and the bottom of the scope.

Co-Authored-By: Joe Brennan <jwsbrennan@gmail.com>